Crate wasmer_vnet

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

BytesMut

A unique reference to a contiguous slice of memory.

HttpStatus

Represents the final result of a HTTP request

IpCidr

Represents an IP address and its netmask

IpRoute

Represents a routing entry in the routing table of the interface

SocketHttpRequest

Holds the interface used to work with a pending HTTP request

SocketReceive

SocketReceiveFrom

UnsupportedVirtualNetworking

Enums

NetworkError

SocketStatus

StreamSecurity

TimeType

Traits

VirtualConnectedSocket

Connected sockets have a persistent connection to a remote peer

VirtualConnectionlessSocket

Connectionless sockets are able to send and receive datagrams and stream bytes to multiple addresses at the same time (peer-to-peer)

VirtualIcmpSocket

ICMP sockets are low level devices bound to a specific address that can send and receive ICMP packets

VirtualNetworking

An implementation of virtual networking

VirtualRawSocket

VirtualSocket

VirtualTcpListener

VirtualTcpSocket

VirtualUdpSocket

VirtualWebSocket

Interface used for sending and receiving data from a web socket

Functions

io_err_into_net_error

net_error_into_io_err

Type Definitions

Result

SocketDescriptor

Socket descriptors are also file descriptors and so all file operations can also be used on sockets