Trait wasmer_vnet::VirtualConnectionlessSocket

pub trait VirtualConnectionlessSocket: VirtualSocket + Debug + Send + Sync + 'static {
    fn send_to(&mut self, data: Bytes, addr: SocketAddr) -> Result<usize>;
    fn recv_from(&mut self) -> Result<SocketReceiveFrom>;
    fn peek_from(&mut self) -> Result<SocketReceiveFrom>;
}

Connectionless sockets are able to send and receive datagrams and stream bytes to multiple addresses at the same time (peer-to-peer)

Required Methods

fn send_to(&mut self, data: Bytes, addr: SocketAddr) -> Result<usize>

Sends out a datagram or stream of bytes on this socket to a specific address

fn recv_from(&mut self) -> Result<SocketReceiveFrom>

Recv a packet from the socket

fn peek_from(&mut self) -> Result<SocketReceiveFrom>

Peeks for a packet from the socket