Trait wasmer_vnet::VirtualRawSocket

pub trait VirtualRawSocket: VirtualSocket + Debug + Send + Sync + 'static {
    fn poll_send(
        &mut self,
        cx: &mut Context<'_>,
        data: &[u8]
    ) -> Poll<Result<usize>>;
    fn try_send(&mut self, data: &[u8]) -> Result<usize>;
    fn poll_flush(&mut self, cx: &mut Context<'_>) -> Poll<Result<()>>;
    fn try_flush(&mut self) -> Result<()>;
    fn poll_recv<'a>(
        &mut self,
        cx: &mut Context<'_>,
        buf: &'a mut [MaybeUninit<u8>]
    ) -> Poll<Result<usize>>;
    fn try_recv(&mut self, buf: &mut [MaybeUninit<u8>]) -> Result<usize>;
    fn set_promiscuous(&mut self, promiscuous: bool) -> Result<()>;
    fn promiscuous(&self) -> Result<bool>;
}

Required Methods

fn poll_send(&mut self, cx: &mut Context<'_>, data: &[u8]) -> Poll<Result<usize>>

Sends out a datagram or stream of bytes on this socket

fn try_send(&mut self, data: &[u8]) -> Result<usize>

Sends out a datagram or stream of bytes on this socket

fn poll_flush(&mut self, cx: &mut Context<'_>) -> Poll<Result<()>>

Attempts to flush the object, ensuring that any buffered data reach their destination.

fn try_flush(&mut self) -> Result<()>

Attempts to flush the object, ensuring that any buffered data reach their destination.

fn poll_recv<'a>(
&mut self,
cx: &mut Context<'_>,
buf: &'a mut [MaybeUninit<u8>]
) -> Poll<Result<usize>>

Recv a packet from the socket

fn try_recv(&mut self, buf: &mut [MaybeUninit<u8>]) -> Result<usize>

Recv a packet from the socket

fn set_promiscuous(&mut self, promiscuous: bool) -> Result<()>

Tells the raw socket and its backing switch that all packets should be received by this socket even if they are not destined for this device

fn promiscuous(&self) -> Result<bool>

Returns if the socket is running in promiscuous mode whereby it will receive all packets even if they are not destined for the local interface