Trait wasmer_vnet::VirtualSocket

pub trait VirtualSocket: Debug + Send + Sync + 'static {
    fn set_ttl(&mut self, ttl: u32) -> Result<()>;
    fn ttl(&self) -> Result<u32>;
    fn addr_local(&self) -> Result<SocketAddr>;
    fn status(&self) -> Result<SocketStatus>;
    fn poll_read_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>;
    fn poll_write_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>;
}

Required Methods

fn set_ttl(&mut self, ttl: u32) -> Result<()>

Sets how many network hops the packets are permitted for new connections

fn ttl(&self) -> Result<u32>

Returns the maximum number of network hops before packets are dropped

fn addr_local(&self) -> Result<SocketAddr>

Returns the local address for this socket

fn status(&self) -> Result<SocketStatus>

Returns the status/state of the socket

fn poll_read_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>

Polls the socket for when there is data to be received

fn poll_write_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>

Polls the socket for when the backpressure allows for writing to the socket